+++ src/components/Scene1.vue
@@ -7,77 +7,17 @@
 </template>
 
 <script>
-//  インポートします。
+//  インポート
 import Vue from 'vue'
+import {NickNameHeader, InputText, InputButton} from './forms/Headers.js'
 
-//  データストア
-var PropertyStore = {
-  debug: true,
-  state: {
-    property: {
-      type: '1',
-      nickname: ''
-    }
-  },
-  setNickNameAction (newValue) {
-    this.debug && console.log('setMessageAction triggered with', newValue)
-    this.state.props.nickname = newValue
-  },
-  clearNickNameAction () {
-    this.debug && console.log('clearMessageAction triggered')
-    this.state.nickname = ''
-  }
-}
-
-//  デフォルト値定義
-export default {
-  data () {
-    return {
-
-    }
-  }
-}
-
-//  ニックネーム表示用のヘッダーコンポーネントを作成
-var NickNameHeader = Vue.extend({
-  name: 'nickname_header',
-  template: '<div class="nickname">{{ sharedState.property.nickname }}</div>',
-  data: function () {
-    return {
-      privateState: {},
-      sharedState: PropertyStore.state
-    }
-  }
-})
 //  コンポーネントを登録
 Vue.component('nickname-header', NickNameHeader)
-
-//  インプットフォーム用のコンポーネントを作成
-var InputText = Vue.extend({
-  name: 'nickname_input',
-  template: '<input class="input" v-model="sharedState.property.nickname" placeholder="NICKNAMEを入力してください。">',
-  data: function () {
-    return {
-      privateState: {},
-      sharedState: PropertyStore.state
-    }
-  }
-})
 //  コンポーネントを登録
 Vue.component('input-text', InputText)
-
-//  ボタン用のコンポーネントを作成
-var InputButton = Vue.extend({
-  name: 'input-button',
-  template: '<button class="button" v-on:click="validate">START</button>',
-  methods: {
-    validate: function (event) {
-      console.log(this.$el)
-    }
-  }
-})
 //  コンポーネントを登録
 Vue.component('input-button', InputButton)
+
 </script>
 
 <!-- Add "scoped" attribute to limit CSS to this component only -->
@@ -104,6 +44,7 @@ Vue.component('input-button', InputButton)
   position: relative;
   margin: 0 auto;
   width: 70pt;
+  background-color: white;
   border: solid 1px silver;
   border-radius: 0.5rem 0.5rem;
   padding: 0.5rem 1.5rem;
